requirements.txt
Flask
rembg
requests
Pillow
